pyproject.toml
[tool.poetry]
name = "tf"
version = "0.1.0"
description = "Python Terraform Provider framework"
authors = ["Hunter Fernandes <hunter@hfernandes.com>"]
license = "MIT"
readme = "README.md"
keywords = ["opentofu", "terraform", "provider", "python"]
homepage = "https://github.com/hfern/tf"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python",
  "Topic :: Software Development :: Libraries :: Python Modules",
]


[tool.poetry.dependencies]
python = "^3.11"
msgpack = "^1.1.0"
cryptography = "^43.0.3"
grpcio = "^1.67.1"


[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.67.1"
ruff = "^0.7.2"
coverage = "^7.6.4"
pyre-check = "^0.9.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
target-version = "py311"
exclude = ["tf/gen"]


[tool.ruff.lint]
extend-select = ["I"]


[tool.coverage.run]
branch = true
omit = [
    "tf/gen/*",
    "*/tests/*",
]

[tool.coverage.report]
show_missing = true
include = [
    "tf/*",
]
